added files for classes and utility functions to reduce the main tool's number of lines
